requirements.txt
streamlit==1.31.0
pandas>=1.3.0
numpy>=1.21.0
google-generativeai>=0.3.0
docx2txt>=0.8
google-genai>=0.3.0
httpx>=0.27.0
